+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.1.2
+* feat: add prefix parameter to customize the golden file name
 * feat: add title parameter to customize the golden file name
 
 ## 1.1.1